gen: Move the public API generator to the pub.rs module
