+++ src/gen/mod.rs
@@ -10,6 +10,7 @@ use syn::Ident;
 mod glib_utils;
 mod toplevel_imports;
 // mod imp;
+// mod pub;
 
 macro_rules! quote_in {
     ($tokens:expr, $($t:tt)*) => {
@@ -160,29 +161,6 @@ impl<'ast> ClassContext<'ast> {
 
     fn get_type_fn_name(&self) -> Ident {
         self.exported_fn_name("get_type")
-    }
-
-    fn pub_impl(&self) -> Tokens {
-        let InstanceName = self.InstanceName;
-        let pub_new_method = self.pub_new_method();
-
-        quote! {
-            impl #InstanceName {
-                #pub_new_method
-            }
-        }
-    }
-
-    fn pub_new_method(&self) -> Tokens {
-        let InstanceName = self.InstanceName;
-        let imp_new_fn_name = self.imp_new_fn_name();
-
-        // FIXME: we should take construct-only arguments and other convenient args to new()
-        quote! {
-            pub fn new() -> #InstanceName {
-                unsafe { from_glib_full(imp::#imp_new_fn_name(/* FIXME: args */)) }
-            }
-        }
     }
 
     fn imp_new_fn_name(&self) -> Ident {
